+++ packages/video_player/lib/video_player.dart
@@ -19,6 +19,7 @@ export 'package:video_player_platform_interface/video_player_platform_interface.
     show
         DataSourceType,
         DurationRange,
+        VideoDrmConfiguration,
         VideoFormat,
         VideoPlayerOptions,
         VideoPlayerWebOptions,
@@ -94,8 +95,16 @@ class VideoAudioTrack {
   }
 
   @override
-  int get hashCode =>
-      Object.hash(id, label, language, isSelected, bitrate, sampleRate, channelCount, codec);
+  int get hashCode => Object.hash(
+    id,
+    label,
+    language,
+    isSelected,
+    bitrate,
+    sampleRate,
+    channelCount,
+    codec,
+  );
 
   @override
   String toString() =>
@@ -114,7 +123,9 @@ class VideoAudioTrack {
 ///
 /// This internal method is used to decouple the public API from the
 /// platform interface implementation.
-VideoAudioTrack _convertPlatformAudioTrack(platform_interface.VideoAudioTrack platformTrack) {
+VideoAudioTrack _convertPlatformAudioTrack(
+  platform_interface.VideoAudioTrack platformTrack,
+) {
   return VideoAudioTrack(
     id: platformTrack.id,
     label: platformTrack.label,
@@ -168,11 +179,16 @@ class VideoPlayerValue {
   });
 
   /// Returns an instance for a video that hasn't been loaded.
-  const VideoPlayerValue.uninitialized() : this(duration: Duration.zero, isInitialized: false);
+  const VideoPlayerValue.uninitialized()
+    : this(duration: Duration.zero, isInitialized: false);
 
   /// Returns an instance with the given [errorDescription].
   const VideoPlayerValue.erroneous(String errorDescription)
-    : this(duration: Duration.zero, isInitialized: false, errorDescription: errorDescription);
+    : this(
+        duration: Duration.zero,
+        isInitialized: false,
+        errorDescription: errorDescription,
+      );
 
   /// This constant is just to indicate that parameter is not passed to [copyWith]
   /// workaround for this issue https://github.com/dart-lang/language/issues/2009
@@ -310,7 +326,8 @@ class VideoPlayerValue {
           : this.errorDescription,
       isCompleted: isCompleted ?? this.isCompleted,
       preventsDisplaySleepDuringVideoPlayback:
-          preventsDisplaySleepDuringVideoPlayback ?? this.preventsDisplaySleepDuringVideoPlayback,
+          preventsDisplaySleepDuringVideoPlayback ??
+          this.preventsDisplaySleepDuringVideoPlayback,
     );
   }
 
@@ -356,7 +373,8 @@ class VideoPlayerValue {
           rotationCorrection == other.rotationCorrection &&
           isInitialized == other.isInitialized &&
           isCompleted == other.isCompleted &&
-          preventsDisplaySleepDuringVideoPlayback == other.preventsDisplaySleepDuringVideoPlayback;
+          preventsDisplaySleepDuringVideoPlayback ==
+              other.preventsDisplaySleepDuringVideoPlayback;
 
   @override
   int get hashCode => Object.hash(
@@ -410,12 +428,14 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
        dataSourceType = platform_interface.DataSourceType.asset,
        formatHint = null,
        httpHeaders = const <String, String>{},
+       drmConfiguration = null,
        isLive = false,
        super(
          VideoPlayerValue(
            duration: Duration.zero,
            preventsDisplaySleepDuringVideoPlayback:
-               videoPlayerOptions?.preventsDisplaySleepDuringVideoPlayback ?? true,
+               videoPlayerOptions?.preventsDisplaySleepDuringVideoPlayback ??
+               true,
          ),
        );
 
@@ -439,6 +459,7 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
     Future<ClosedCaptionFile>? closedCaptionFile,
     this.videoPlayerOptions,
     this.httpHeaders = const <String, String>{},
+    this.drmConfiguration,
     this.viewType = platform_interface.VideoViewType.textureView,
     this.isLive = false,
   }) : _closedCaptionFileFuture = closedCaptionFile,
@@ -448,7 +469,8 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
          VideoPlayerValue(
            duration: Duration.zero,
            preventsDisplaySleepDuringVideoPlayback:
-               videoPlayerOptions?.preventsDisplaySleepDuringVideoPlayback ?? true,
+               videoPlayerOptions?.preventsDisplaySleepDuringVideoPlayback ??
+               true,
          ),
        );
 
@@ -461,12 +483,17 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
   ///
   /// [httpHeaders] option allows to specify HTTP headers
   /// for the request to the [dataSource].
+  ///
+  /// The [drmConfiguration] option allows playing DRM-protected content. It
+  /// must be a configuration supported by the current platform implementation.
+  /// DRM is supported only for network sources.
   VideoPlayerController.networkUrl(
     Uri url, {
     this.formatHint,
     Future<ClosedCaptionFile>? closedCaptionFile,
     this.videoPlayerOptions,
     this.httpHeaders = const <String, String>{},
+    this.drmConfiguration,
     this.viewType = platform_interface.VideoViewType.textureView,
     this.isLive = false,
   }) : _closedCaptionFileFuture = closedCaptionFile,
@@ -477,7 +504,8 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
          VideoPlayerValue(
            duration: Duration.zero,
            preventsDisplaySleepDuringVideoPlayback:
-               videoPlayerOptions?.preventsDisplaySleepDuringVideoPlayback ?? true,
+               videoPlayerOptions?.preventsDisplaySleepDuringVideoPlayback ??
+               true,
          ),
        );
 
@@ -496,12 +524,14 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
        dataSourceType = platform_interface.DataSourceType.file,
        package = null,
        formatHint = null,
+       drmConfiguration = null,
        isLive = false,
        super(
          VideoPlayerValue(
            duration: Duration.zero,
            preventsDisplaySleepDuringVideoPlayback:
-               videoPlayerOptions?.preventsDisplaySleepDuringVideoPlayback ?? true,
+               videoPlayerOptions?.preventsDisplaySleepDuringVideoPlayback ??
+               true,
          ),
        );
 
@@ -524,12 +554,14 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
        package = null,
        formatHint = null,
        httpHeaders = const <String, String>{},
+       drmConfiguration = null,
        isLive = false,
        super(
          VideoPlayerValue(
            duration: Duration.zero,
            preventsDisplaySleepDuringVideoPlayback:
-               videoPlayerOptions?.preventsDisplaySleepDuringVideoPlayback ?? true,
+               videoPlayerOptions?.preventsDisplaySleepDuringVideoPlayback ??
+               true,
          ),
        );
 
@@ -545,6 +577,12 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
   /// **Android only**. Will override the platform's generic file format
   /// detection with whatever is set here.
   final platform_interface.VideoFormat? formatHint;
+
+  /// The DRM configuration for the video, if any.
+  ///
+  /// Only set for network sources; it is always null for asset, file, and
+  /// content URI sources.
+  final platform_interface.VideoDrmConfiguration? drmConfiguration;
 
   /// Describes the type of data source this [VideoPlayerController]
   /// is constructed with.
@@ -590,7 +628,8 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
 
   /// Attempts to open the given [dataSource] and load metadata about the video.
   Future<void> initialize() async {
-    final bool allowBackgroundPlayback = videoPlayerOptions?.allowBackgroundPlayback ?? false;
+    final bool allowBackgroundPlayback =
+        videoPlayerOptions?.allowBackgroundPlayback ?? false;
     if (!allowBackgroundPlayback) {
       _lifeCycleObserver = _VideoAppLifeCycleObserver(this);
     }
@@ -611,6 +650,7 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
           uri: dataSource,
           formatHint: formatHint,
           httpHeaders: httpHeaders,
+          drmConfiguration: drmConfiguration,
         );
       case platform_interface.DataSourceType.file:
         dataSourceDescription = platform_interface.DataSource(
@@ -632,11 +672,14 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
     );
 
     if (videoPlayerOptions?.mixWithOthers != null) {
-      await _videoPlayerPlatform.setMixWithOthers(videoPlayerOptions!.mixWithOthers);
+      await _videoPlayerPlatform.setMixWithOthers(
+        videoPlayerOptions!.mixWithOthers,
+      );
     }
 
     _playerId =
-        (await _videoPlayerPlatform.createWithOptions(creationOptions)) ?? kUninitializedPlayerId;
+        (await _videoPlayerPlatform.createWithOptions(creationOptions)) ??
+        kUninitializedPlayerId;
     _creatingCompleter!.complete(null);
     final initializingCompleter = Completer<void>();
 
@@ -647,7 +690,10 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
 
     // Apply the web-specific options
     if (kIsWeb && videoPlayerOptions?.webOptions != null) {
-      await _videoPlayerPlatform.setWebOptions(_playerId, videoPlayerOptions!.webOptions!);
+      await _videoPlayerPlatform.setWebOptions(
+        _playerId,
+        videoPlayerOptions!.webOptions!,
+      );
     }
 
     void eventListener(platform_interface.VideoEvent event) {
@@ -697,7 +743,10 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
           value = value.copyWith(isBuffering: false);
         case platform_interface.VideoEventType.isPlayingStateUpdate:
           if (event.isPlaying ?? false) {
-            value = value.copyWith(isPlaying: event.isPlaying, isCompleted: false);
+            value = value.copyWith(
+              isPlaying: event.isPlaying,
+              isCompleted: false,
+            );
           } else {
             value = value.copyWith(isPlaying: event.isPlaying);
           }
@@ -774,7 +823,8 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
     bool preventsDisplaySleepDuringVideoPlayback,
   ) async {
     value = value.copyWith(
-      preventsDisplaySleepDuringVideoPlayback: preventsDisplaySleepDuringVideoPlayback,
+      preventsDisplaySleepDuringVideoPlayback:
+          preventsDisplaySleepDuringVideoPlayback,
     );
     await _applyPreventsDisplaySleepDuringVideoPlayback();
   }
@@ -810,7 +860,9 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
       await _videoPlayerPlatform.play(_playerId);
 
       _timer?.cancel();
-      _timer = Timer.periodic(const Duration(milliseconds: 100), (Timer timer) async {
+      _timer = Timer.periodic(const Duration(milliseconds: 100), (
+        Timer timer,
+      ) async {
         if (_isDisposed) {
           return;
         }
@@ -907,7 +959,10 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
   ///   the plugin also reports errors.
   Future<void> setPlaybackSpeed(double speed) async {
     if (speed < 0) {
-      throw ArgumentError.value(speed, 'Negative playback speeds are generally unsupported.');
+      throw ArgumentError.value(
+        speed,
+        'Negative playback speeds are generally unsupported.',
+      );
     } else if (speed == 0) {
       throw ArgumentError.value(
         speed,
@@ -929,7 +984,10 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
   /// * >0: The caption will have a negative offset. So you will get caption text from the past.
   /// * <0: The caption will have a positive offset. So you will get caption text from the future.
   void setCaptionOffset(Duration offset) {
-    value = value.copyWith(captionOffset: offset, caption: _getCaptionAt(value.position));
+    value = value.copyWith(
+      captionOffset: offset,
+      caption: _getCaptionAt(value.position),
+    );
   }
 
   /// The closed caption based on the current [position] in the video.
@@ -950,7 +1008,12 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
 
     final int captionIndex = collection.binarySearch<Caption>(
       sortedCaptions,
-      Caption(number: -1, start: delayedPosition, end: delayedPosition, text: ''),
+      Caption(
+        number: -1,
+        start: delayedPosition,
+        end: delayedPosition,
+        text: '',
+      ),
       compare: (Caption candidate, Caption search) {
         if (search.start < candidate.start) {
           return 1;
@@ -979,14 +1042,18 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
   /// Sets a closed caption file.
   ///
   /// If [closedCaptionFile] is null, closed captions will be removed.
-  Future<void> setClosedCaptionFile(Future<ClosedCaptionFile>? closedCaptionFile) async {
+  Future<void> setClosedCaptionFile(
+    Future<ClosedCaptionFile>? closedCaptionFile,
+  ) async {
     _closedCaptionFileFuture = closedCaptionFile;
     // Reset sorted captions to force re-sort when setting a new file
     _sortedCaptions = null;
     await _updateClosedCaptionWithFuture(closedCaptionFile);
   }
 
-  Future<void> _updateClosedCaptionWithFuture(Future<ClosedCaptionFile>? closedCaptionFile) async {
+  Future<void> _updateClosedCaptionWithFuture(
+    Future<ClosedCaptionFile>? closedCaptionFile,
+  ) async {
     if (closedCaptionFile != null) {
       _closedCaptionFile = await closedCaptionFile;
 
@@ -1041,8 +1108,8 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
     if (!value.isInitialized) {
       return <VideoAudioTrack>[];
     }
-    final List<platform_interface.VideoAudioTrack> platformTracks = await _videoPlayerPlatform
-        .getAudioTracks(_playerId);
+    final List<platform_interface.VideoAudioTrack> platformTracks =
+        await _videoPlayerPlatform.getAudioTracks(_playerId);
     return platformTracks.map(_convertPlatformAudioTrack).toList();
   }
 
@@ -1106,10 +1173,13 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
     if (_isDisposedOrNotInitialized) {
       return <VideoTrack>[];
     }
-    final List<platform_interface.VideoTrack> platformTracks = await _videoPlayerPlatform
-        .getVideoTracks(_playerId);
+    final List<platform_interface.VideoTrack> platformTracks =
+        await _videoPlayerPlatform.getVideoTracks(_playerId);
     return platformTracks
-        .map((platform_interface.VideoTrack track) => VideoTrack._fromPlatform(track))
+        .map(
+          (platform_interface.VideoTrack track) =>
+              VideoTrack._fromPlatform(track),
+        )
         .toList();
   }
 
@@ -1302,7 +1372,11 @@ class VideoScrubber extends StatefulWidget {
   ///
   /// [controller] is the [VideoPlayerController] that will be controlled by
   /// this scrubber.
-  const VideoScrubber({super.key, required this.child, required this.controller});
+  const VideoScrubber({
+    super.key,
+    required this.child,
+    required this.controller,
+  });
 
   /// The widget that will be displayed inside the gesture detector.
   final Widget child;
@@ -1348,7 +1422,8 @@ class _VideoScrubberState extends State<VideoScrubber> {
         seekToRelativePosition(details.globalPosition);
       },
       onHorizontalDragEnd: (DragEndDetails details) {
-        if (_controllerWasPlaying && controller.value.position != controller.value.duration) {
+        if (_controllerWasPlaying &&
+            controller.value.position != controller.value.duration) {
           controller.play();
         }
       },
@@ -1442,7 +1517,10 @@ class _VideoProgressIndicatorState extends State<VideoProgressIndicator> {
       final double maxBuffering = duration == 0.0
           ? 0.0
           : controller.value.buffered
-                    .map((platform_interface.DurationRange range) => range.end.inMilliseconds)
+                    .map(
+                      (platform_interface.DurationRange range) =>
+                          range.end.inMilliseconds,
+                    )
                     .fold(0, math.max) /
                 duration;
       progressIndicator = Stack(
@@ -1471,7 +1549,10 @@ class _VideoProgressIndicatorState extends State<VideoProgressIndicator> {
       child: progressIndicator,
     );
     if (widget.allowScrubbing) {
-      return VideoScrubber(controller: controller, child: paddedProgressIndicator);
+      return VideoScrubber(
+        controller: controller,
+        child: paddedProgressIndicator,
+      );
     } else {
       return paddedProgressIndicator;
     }
@@ -1523,7 +1604,9 @@ class ClosedCaption extends StatelessWidget {
 
     final TextStyle effectiveTextStyle =
         textStyle ??
-        DefaultTextStyle.of(context).style.copyWith(fontSize: 36.0, color: Colors.white);
+        DefaultTextStyle.of(
+          context,
+        ).style.copyWith(fontSize: 36.0, color: Colors.white);
 
     return Align(
       alignment: Alignment.bottomCenter,
@@ -1634,7 +1717,16 @@ class VideoTrack {
   }
 
   @override
-  int get hashCode => Object.hash(id, isSelected, label, bitrate, width, height, frameRate, codec);
+  int get hashCode => Object.hash(
+    id,
+    isSelected,
+    label,
+    bitrate,
+    width,
+    height,
+    frameRate,
+    codec,
+  );
 
   @override
   String toString() =>