+++ packages/video_player_android/lib/video_player_android.dart
@@ -1,0 +1,6 @@
+// Copyright 2013 The Flutter Authors
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+export 'src/android_video_player.dart';
+export 'src/widevine_drm_configuration.dart';